+++ client/src/App.jsx
@@ -5,7 +5,7 @@ import Footer from './components/Footer.jsx';
 import Header from './components/Header.jsx';
 import Login from './pages/Login.jsx';
 // import Login from './pages/Login.jsx';
-// import SignUp from './pages/SignUp.jsx';
+import SignUp from './pages/SignUp.jsx';
 
 function App() {
   return (
